refactor(frontend): split setup components
