+++ new/src/frontend/acc2/src/components/setup/parameters/publication.tsx
@@ -1,0 +1,12 @@
+import { Parameters } from "@acc2/api/parameters/types";
+import { Publication } from "../publication";
+
+export type ParametersPublicationProps = {
+  parameters?: Parameters;
+};
+
+export const ParametersPublication = ({
+  parameters,
+}: ParametersPublicationProps) => {
+  return <>{parameters && <Publication publicationSource={parameters} />}</>;
+};
